moved FriendListDetails to its own component
