+++ friends/src/Components/FriendList.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 // import { Link } from 'react-router-dom'
 import axios from 'axios';
 import AddFriendForm from './AddFriendForm';
+import FriendDetails from './FriendDetails';
 
 class FriendList extends Component {
   constructor(props) {
@@ -56,24 +57,6 @@ class FriendList extends Component {
     );
   }
 
-}
-
-function FriendDetails({ friend }) {
-  const { name, age, email } = friend;
-  return (
-    <div className="friend-card">
-      <h2>{name}</h2>
-      <div className="name">
-        Name: <em>{name}</em>
-      </div>
-      <div className="age">
-        Age: <strong>{age}</strong>
-      </div>
-      <div className="email">
-        Email: <strong>{email}</strong>
-      </div>
-    </div>
-  );
 }
 
 export default FriendList;
